+++ cloudflare-access-sso.php
@@ -35,13 +35,15 @@ defined( 'ABSPATH' ) || exit;
 
 // The Cloudflare Team Name is required
 if ( ! defined( 'CF_ACCESS_TEAM_NAME' ) ) {
-	error_log( 'Cloudflare Access SSO Error: CF_ACCESS_TEAM_NAME is not defined.' ); // phpcs:ignore WordPress.PHP.DevelopmentFunctions.error_log_error_log
+	error_log( 'Cloudflare Access SSO Error: Required constant: CF_ACCESS_TEAM_NAME is not defined.' ); // phpcs:ignore WordPress.PHP.DevelopmentFunctions.error_log_error_log
+	error_log( sprintf( 'Review the documentation at %s for setup instructions.', CF_ACCESS_GITHUB_URL ) ); // phpcs:ignore WordPress.PHP.DevelopmentFunctions.error_log_error_log
 	return;
 }
 
 // The Cloudflare Application ID is required
 if ( ! defined( 'CF_ACCESS_AUD' ) ) {
-	error_log( 'Cloudflare Access SSO Error: CF_ACCESS_AUD is not defined.' ); // phpcs:ignore WordPress.PHP.DevelopmentFunctions.error_log_error_log
+	error_log( 'Cloudflare Access SSO Error: Required constant: CF_ACCESS_AUD is not defined.' ); // phpcs:ignore WordPress.PHP.DevelopmentFunctions.error_log_error_log
+	error_log( sprintf( 'Review the documentation at %s for setup instructions.', CF_ACCESS_GITHUB_URL ) ); // phpcs:ignore WordPress.PHP.DevelopmentFunctions.error_log_error_log
 	return;
 }
 
@@ -55,23 +57,34 @@ if ( ! defined( 'CF_ACCESS_LEEWAY' ) ) {
 	define( 'CF_ACCESS_LEEWAY', 60 );
 }
 
+// Default to not creating accounts
+if ( ! defined( 'CF_ACCESS_CREATE_ACCOUNT' ) ) {
+	define( 'CF_ACCESS_CREATE_ACCOUNT', false );
+}
+
+// Default to subscriber role for new accounts
+if ( ! defined( 'CF_ACCESS_NEW_USER_ROLE' ) ) {
+	define( 'CF_ACCESS_NEW_USER_ROLE', 'subscriber' );
+}
+
 // Useful global constants
-define( 'CLOUDFLARE_ACCESS_SSO_PLUGIN_VERSION', '1.0.3' );
-define( 'CLOUDFLARE_ACCESS_SSO_PLUGIN_URL', plugin_dir_url( __FILE__ ) );
-define( 'CLOUDFLARE_ACCESS_SSO_PLUGIN_PATH', plugin_dir_path( __FILE__ ) );
-define( 'CLOUDFLARE_ACCESS_SSO_PLUGIN_INC', CLOUDFLARE_ACCESS_SSO_PLUGIN_PATH . 'includes/' );
+define( 'CF_ACCESS_SSO_PLUGIN_VERSION', '1.0.3' );
+define( 'CF_ACCESS_SSO_PLUGIN_URL', plugin_dir_url( __FILE__ ) );
+define( 'CF_ACCESS_SSO_PLUGIN_PATH', plugin_dir_path( __FILE__ ) );
+define( 'CF_ACCESS_SSO_PLUGIN_INC', CF_ACCESS_SSO_PLUGIN_PATH . 'includes/' );
+define( 'CF_ACCESS_GITHUB_URL', 'https://github.com/jamesmorrison/cloudflare-access-sso' );
 
 // Load plugin classes
-if ( ! file_exists( CLOUDFLARE_ACCESS_SSO_PLUGIN_PATH . 'vendor/autoload.php' ) ) {
+if ( ! file_exists( CF_ACCESS_SSO_PLUGIN_PATH . 'vendor/autoload.php' ) ) {
 	error_log( 'Cloudflare Access SSO Error: Composer dependencies are missing. Please run `composer install` in the plugin directory.' ); // phpcs:ignore WordPress.PHP.DevelopmentFunctions.error_log_error_log
 	return;
 }
 
 // Load composer dependencies
-require_once CLOUDFLARE_ACCESS_SSO_PLUGIN_PATH . 'vendor/autoload.php';
+require_once CF_ACCESS_SSO_PLUGIN_PATH . 'vendor/autoload.php';
 
 // Load plugin files
-require_once CLOUDFLARE_ACCESS_SSO_PLUGIN_INC . '/core.php';
+require_once CF_ACCESS_SSO_PLUGIN_INC . '/core.php';
 
 // Activation / Deactivation
 register_activation_hook( __FILE__, '\CloudflareAccessSSO\Core\activate' );